Fix or silence compilation warnings
Following warnings were fixed/silenced:
-Wsign-conversion
-Wimplicit-int-conversion
-Wnull-pointer-subtraction
-Wunused-variable
-Wtautological-type-limit-compare
-Wtautological-unsigned-enum-zero-compare

+++ websocketpp/http/impl/request.hpp
@@ -89,7 +89,7 @@ inline size_t request::consume(char const * buf, size_t len, lib::error_code & e
 
             // check that the confirmed header bytes plus the outstanding
             // candidate bytes do not put us over the header size limit.
-            if (m_header_bytes + (end - begin) > max_header_size) {
+            if (m_header_bytes + static_cast<size_t>(end - begin) > max_header_size) {
                 ec = error::make_error_code(error::request_header_fields_too_large);
                 return 0;
             }
@@ -113,7 +113,7 @@ inline size_t request::consume(char const * buf, size_t len, lib::error_code & e
         // represents a line to be processed
 
         // update count of header bytes read so far
-        m_header_bytes += (end-begin+sizeof(header_delimiter));
+        m_header_bytes += (static_cast<size_t>(end-begin)+sizeof(header_delimiter));
         
 
         if (m_header_bytes > max_header_size) {
